Make telepathy-integration-daemon a Nepomuk Service
Having a Nepomuk service instead of a daemon allows to check for services
dependencies just by adding 1 line in the .desktop file and the service will be
started as soon as the dependencies are ready.

It also adds automatically a dbus interface that allows to check if the service
is available or running, and to start and stop the service.

See also http://reviewboard.kde.org/r/5449
Closes bug #246247


svn path=/trunk/playground/network/telepathy-integration-daemon/; revision=1181210