Refactor the Nepomuk service to be have a better Model-View-Controller architecture.
This makes it more maintainable, as well as eaiser to extend in the future.
Thanks to Andrunko for providing the outline of this new architecture.

All basic features have been ported, although capabilities and avatars have not been ported yet - they will come later as they are having some major changes.
Also, took the oppurtunity to port to a recent trunk of shared-desktop-ontologies and telepathy-qt4 0.5.1