Make the tests work with the NepomukStorage
NepomukStorage is asynchronous and has an approximate delay of 500 msecs plus
dbus round trip time. This has been done to improve performance by caching
changes and doing them in bulk.

Therefore, we emit a signal graphSaved() in NepomukStorage, when the changes
are saved, and wait for that signal in the tests.

We do not verify the if the graphSaved signal is emitted, cause sometimes the
data already exists and nothing is done.