Make MessageProcessor and KCM use ktelepathyrc
I didn't realize that the MessagesConfig was saving its settings in a
different file than MessageProcessor was loading from.

This promptly fixes that, but I don't like the idea of hardcoding
the file in many places, nor kludging MessageProcessor, I think I'll
make a separate class for manging plugin settings.