NepomukFeeder class
A horrible horrible hack that I just wanted to try out. It seems to
simplify the code a lot, but is NOT the proper way of doing things as it
skips all the checks we spent years building.

The relevant code should probably go in the DMS under a new function -
overWriteProperties or something.