Set all contacts to offline when account goes offline
The presenceChanged(..) signal is never emitted (bug in tp-qt?)
when the account disconnects and so Nepomuk is left with invalid
contact presences. This sets all contacts to offline when
the connection drops. The presence is re-set when the account
is online thanks to the presenceChanged(..) being emitted for all contacts.

Reviewed-by: David Edmundson
REVIEW: 107385