Avoid calling KWallet::openWallet several times
In order to avoid the problem when kwallet is closed while kcm accounts
is open, we introduced a bug: if more than one openWallet call is done
at the same time, we reset the scoped pointer with a new wallet, the
old wallet is deleted and therefore the signal is never emitted.

This patch uses avoids trying to open the wallet several times if it is
already being opened.

Reviewed-by: George Kiagiadakis <george.kiagiadakis@collabora.com>
REVIEW: 107426
BUG: 310505
FIXED-IN: 0.5.2