Allow empty groups in AbstractGroupingProxyModel
AbstractGroupingProxyModel cannot show empty groups as it does not know
about them.

Add two public members forceGroup, unforceGroup, which allow empty groups
to be added to the model. Group will only be removed when empty AND the group
is not enforced.

Alter AccountsTreeProxyModel to force groups to be shown for all accounts