Add a new class for processing outgoing messages
Currently we re-use KTp::Message, this exposes a lot of setters and getters that have no use before we send a message.
It also mixes HTML and plain text which has already led to one bug, and would probably lead to more.

This keeps the API cleaner and simpler

REVIEW: 109403