Restore printing tp-qt version and fix it for debug kdelibs
Commit 6cbefb04441bab0d25dde78d95a2adee4d163fd5 "Prevent crashes with
debug kdelibs" should probably have been reverting commit
58d7eb5a461bc597455065da3a4cd813e4f0a15a "Print tp-qt library name and
version in tp-qt debug"

If the line cannot be printed, it doesn't make sense to keep the extra
string and the extra check.

Anyway the problem seem to come with a Q_ASSERT_X in kdelibs that
searches for "(" (see kdelibs kdecore/io/kdebug.cpp.html:577)

This patch restore the previous behaviour, and should fix the crash.

Also add a ":" to the string, because I realized that everything that
comes before the white space is stripped away, so now it prints
something like "tp-qt:0.9.3.1"

Reviewed-by: David Edmundson <kde@davidedmundson.co.uk>
REVIEW: 109852