Load translation catalog and fix use of i18n in global scope
In order for KLocale to work, i18n() must not be used
in global static variables, because these are initialised
before KLocale. Use I18N_NOOP instead to mark the string
for translation and use i18n to perform the actual
lookup of the translated string only before displaying
the string to user.

Reviewed-By: David Edmundson
BUG: 318190
FIXED-IN: 0.6.1