Make sure PersonContact is a nie:DataObject
That way they do not get identified.

Nepomuk has this resource identification code which tries to find other
PersonContacts which are similar to the one being pushed. That code is
useful in the case when pushing generic stuff like artists, and email
contacts, but not in this case where each PersonContact directly
corresponds to a vcard. We do not want it merged with any other contact.