Create and cache the contact groups
Earlier, if a group contained more than one contact then it the contact
group was created that many times cause a new SimpleResource was created
each time.

With this method we cache the contact groups so that they are only
created once, and once they have been created we do not unnecessarily
send a new SimpleResource to get merged into the existing contact group.

REVIEW: 110739