Upgrade the Tp::Connection to upgrade FeatureRoster inside GlobalContactManager
GlobalContactManager is useless without FeatureRoster so it makes sense
that if one instantiates this class, we upgrade the connection.

This will allow apps that only occasionally need a full contact list
such as contact-applet and text-ui to not rely on FeatureRoster and thus
will can a connection object without loading details on every contact ever.

This is an important step towards a singleton AccountManager, in addition
text-ui will load substantially faster.

REVIEW: 111440