Initial interface with minimal implementation
The idea behind the design is that even the LogManager itself, is
just a "plugin", and the pending operations that the LogManager
returns are special implementations of the base PendingLogger*
operations which only invoke the same operation on every loaded
plugin, then collect results from all the operations and emit
finished().

This eliminates need for additional layers of abstraction that we
would have to introduce otherwise.

In theory, we could even make LogManager to inherit from
AbstractLoggerPlugin.