Don't delete LogsImporter thread while it's still running
When LogsImporter is deleted, it just deletes the importer thread
even when it's running, which will obviously crash. This patch adds
a safety guard (calling QThread::wait() before deleting it) and a
method to stop the thread prematurely so that we don't wait() for
the entire import to finish.

REVIEW: 112188