Set accountUniqueIdentifier on contacts
Tp::Connection has no method to get the holding Tp::AccountPtr

This makes it difficult to work out which account a contact belongs to.
This can't be fixed in TpQt in 0.9.x

Instead create a KTp::Account which sets a property on connection
whenever the connection chanages.

This is exposed from KTp::Contact

REVIEW: 114643