Add a workaround for broken sorting
There is a bug in QSortFilterProxyModel in which if sort() is called
on a model where all items are filtered out it will not start sorting
as it cannot map the proxy column to the source column.

This gets updated whenever there's an insertRows but not from
a dataChanged.

With the kpeople cache we load all items, but
none make it through the filter as everyone is assumed to be
offline whilst we query telepathy over DBus. When we get data
from Tp we will only ever call dataChanged() not insertRows()
so we hit the Qt bug.

This bug is fixed in Qt 5.2.1, but not in 4.x

REVIEW:115828

BUG:327008
FIXED-IN: 0.8.0

DIGEST: Please do not put this in the commit log. No-one should see
how horrible this is.