Added presence request message in the add contact dialog
Added an optional text input line for the capability of sending a message with presence requests.

I'm not entirely sure of the HIG for making it clear that it's an optional field, so put the label below for now - putting 'message (optional)' on the left-hand label looks unreasonably awkward to me.

REVIEW: 117138
BUG: 318529