Remove internal countContacts function
This method counted the contacts in a group every time a contact in that
group changed. This meant on every insertion we would loop through the
entire set of contacts to make the new count.

Assuming all contacts come online we end up making contacts^2 calls to
data() which is obviously really slow. On my system KPeople::data()
showed over 300,000 calls on load

Instead strip this code and re-insert a simple filter at the contact
list level (see patch to contact list)