Map connection presence type to string
In libkpeople we pass the presence around as a string as that's all
KABC::custom() supports.

We used to use presence->status(); this is problematic as it is free
text which can contain anything; especially when on multiple protocols.

A presence might have the status "chat" which is mapped to the presence
type Tp::ConnectionPresenceTypeOnline. We want to use this mapping and
only have a subset of presence strings.

BUG: 334207