Set the Unknown and Error presences to something
Otherwise the translation proxy will filter these out when we're online
(as it rejects contacts without any presence), but these would appear
when the user goes offline because at that point it's fetched from the
cache, which simply contains all contacts and sets all of them to
"offline".

Additionally, we do want to have the blocked contacts available for KTp
so the user can ublock them if needed. Their filtering is done by the
KTp filter proxy and contact list setting.

Issue found & fixed thanks to the testing application in tests/ \o/