Improve the contacts cache for KPeople
Currently when account goes online from offline, the contact cache is
completely nuked and recreated. This is fast but has one drawback - it
will store avatars only of those contacts that are online in the moment
of connection, so basically with each connection the available avatars
change a bit. To compensate, we've been storing the avatar token in a
config file, but that's used only for KTp::Contact.

Now with KPeople, if the account is offline, we don't have access to
KTp::Contact and have avatars only from the last cache recreation (that
means that when you go offline, half of your contacts suddenly loose
avatars). This patch now reloads the cache for the account that went
offline. But as the contact may have not came online during the last
connection, it now also loads the avatars from the config file if the
database has nothing (this matches the KTp::Contact behavior).

This is far from ideal as both caches should be merged into just one,
but given the time constraint, this will have to do for now.

REVIEW: 122444