Add simple accounts migration
This will be run after the kded module loads the plugin. After
connecting for the first time, the user will be asked for password and
if he chooses to remember the password, it will become stored in
KAccounts.

If an account of unknown type (unknown to KAccounts) is found, it will
use a generic fallback provider and service.