Add support for "Connect and start chat" action into kpeople actions
When the contact is offline, special action is returned which will first
connect the account and then start the chat.

This is done for text chat only as there is no way of knowing if the
target contact has a support for audio/video calls before it's
connected, while text chats will work in 98% of the cases.

REVIEW: 123185