Delay the initialization of the accountUID a bit
Text-ui creates contacts in the channel before the
Tp::Account::connectionChanged is signaled and the
accountUID property set, so this delays it to the
point where it is actually needed.

Proper fix should go to TelepathyQt to emit the signals
in proper order, but that would be very complex patch.
So this will have to do for now.

Reviewed-by: David Edmundson