Relicense invalid license to LGPL 2.1+
Scott Kitterman noticed the license for this code was
"GPL 2.1+" which doesn't exist.

I (Diane Trout) tried to contact the author on a few occasions but
did not see a response.

I believe this code was created as a Google Summer of Code project and
so the following criteria which states that the student needs to choose
a license approved by the mentoring organization.

http://www.google-melange.com/gsoc/document/show/gsoc_program/google/gsoc2015/help_page?PageSpeed=noscript#2._What_licenses_do_I_have_to_choose

The most likely thing is that the original license block is a
typo and the student had first written GPL-3, but then changed it to
GPL 2.1, but forgot to add the "Lesser".

This is some additional files that I missed from my previous
review: 125780 but the same justifications should apply.