[declarative] Throw out QSqlQueryModel and use QSqlQuery directly instead
The problem with QSqlQueryModel is that it is a static model in a sense
that whenever the database updates, the model needs to be /fully/
reloaded, which makes adding only new stuff impossible.

So it had to go and it was replaced by couple queries instead.
Everything works just fine.