Fix escape/url filter URL placeholder backward substitution
The placeholder ends with a decimal URL index. The substitution is an
incremental pass from zero index, so when the placeholder is substituted
back to URL the replacing code matches incomplete index and we got a
messed URLs.

(Replace of `#K#T#P1` messes up `#K#T#P10` and `#K#T#P11`)

Add a end marker to the placeholder scheme to never mess URL.

BUG: 384968
FIXED-IN: 18.12.3

Differential Revision: https://phabricator.kde.org/D19477